Create com-1
This is my first github commit so bear with me! 

I am an avid stock market options enthusiast and aspiring data scientist. This script is to find unusual option activity in the stock market. 

In this script, we leverage the vast stock market data from barchart.com which provides in depth option activity for a reasonable price to their premier members. ($30/month)

Once on the home page, we navigate to 'Options' tab, then in the drop down we select 'Unusual Options Activity'.

From there, we can put the path to the csv file in the  'df = pd.read_csv('your_new_csv.csv')' which will be read into a pandas dataframe.

Once the information has been loaded in via csv, you are ready to run the script and explore some unusual options activity!

If you would like to change some things or added your own, I've left some filters there which may be suitable for your trading interests.

I'm currently working on building a portfolio for a career opportunity in data science so any help would be greatly appreciated. :)

Best,
Jacob